New pipeline to add changes that align with assignment requirements. Assignment has the following requirements regarding tokenisation : 1. Decide for a tokenizer (BPE in our case) 2. Implement the tokenizer you decided on in a new code cell in the Jupyter notebook. IMPORTANT: Do not delete / overwrite the existing simple tokenization method (which is white space) as you will need it later on! c) Tokenize a few sample texts with the decided tokenizer and decide whether further text normalization like stemming or custom regular expressions is needed. Add at least two tokenized texts (Given accuracy is lower and issues in the text...further normalization is decided) -Add the text-normalization steps you decided on to your tokenization function and show how the normalization steps improve the tokenization compared to raw tokenization. -Compare accuracy on the test set of the pipeline using the default white space tokenization vs your chosen improved tokenization without additional text normalization (BPE alone) vs your chosen improved tokenization with your additional text normalization (BPE plus normalization steps)
Given these requirements above, I've started a new one notebook with different pipeline with all the steps required and took out all the preprocessing steps I did before tokenizing in the other notebook.

If this is ok, we might add the new code cell implementing a TF-IDF embedding and comparing accuracy on the test set of the pipeline using the default Bag-of-Words embedding vs the TF-IDF-Embedding you added as per assignment requirements.

Could also add NN code you did in the other notebook